net: dsa: yt921x: Abuse SGMII for REVSGMII
Although it shows that YT9215s + IPQ5018 works on 2500Base-X mode in an
old unofficial port version of OpenWRT, it stops working in the latest
version, probably due to unknown changes in qca-ssdk (see link below).
REVSGMII (SGMII in PHY role) is the only available mode for (the
mainline) OpenWRT, however we lack PHY_INTERFACE_MODE_REVSGMII in the
linux kernel, so we abuse SGMII to denote REVSGMII for the moment, until
a solution (either introducing REVSGMII mode or fixing qca-ssdk) is
found.

Link: https://forum.openwrt.org/t/241325

Signed-off-by: David Yang <mmyangfl@gmail.com>